Add files via upload
Browser exploitation framework is an open source tool in Kali Linux that we used to prepare and demonstrate how BeEF can be used. This research, demonstration and presentation were carried out under the supervision of university of Adelaide cybersecurity Bootcamp. All testing was done in a secure lab environment and no real companies or systems were used. This research presentation and its demonstration are only for educational purposes.